move subprocess to gateways
